Update NetSalary.c
I added the option to calculate the tax percentage and subtract that amount from the net salary. This gives you the final salary after tax. 
I also made the option to enter the over time rate instead of picking an option from the list. 
At the end, the overtime hours are multiplied by the overtime rate. The answer is then added to the salary to get the net salary. 
The tax amount is calculated by dividing the tax % by 100 and then multiplying to the net salary. 
The final salary is the net salary minus the tax amount. 
At the end, the net salary is displayed and then the final salary after tax is also displayed.